Minor style fixes and refactoring
