+++ modules/magnolia/src/main/scala/pureconfig/module/magnolia/auto/reader.scala
@@ -1,4 +1,4 @@
-package pureconfig.module.magnolia.reader
+package pureconfig.module.magnolia.auto
 
 import scala.language.experimental.macros
 
@@ -8,10 +8,10 @@ import pureconfig.generic.{ CoproductHint, ProductHint }
 import pureconfig.module.magnolia.MagnoliaConfigReader
 
 /**
- * An object that, when imported, provides methods for deriving `ConfigReader` instances on demand for value classes,
- * tuples, case classes and sealed traits. The generation of `ConfigReader`s is done by Magnolia.
+ * An object that, when imported, provides implicit `ConfigReader` instances for value classes, tuples, case classes and
+ * sealed traits. The generation of `ConfigReader`s is done by Magnolia.
  */
-object semiauto {
+object reader {
   type Typeclass[A] = ConfigReader[A]
 
   def combine[A: ProductHint](ctx: CaseClass[ConfigReader, A]): ConfigReader[A] =
@@ -20,5 +20,5 @@ object semiauto {
   def dispatch[A: CoproductHint](ctx: SealedTrait[ConfigReader, A]): ConfigReader[A] =
     MagnoliaConfigReader.dispatch(ctx)
 
-  def deriveReader[A]: ConfigReader[A] = macro Magnolia.gen[A]
+  implicit def exportReader[A]: ConfigReader[A] = macro Magnolia.gen[A]
 }
