+++ modules/magnolia/src/test/scala/pureconfig/module/magnolia/CoproductConvertersSuite.scala
@@ -6,8 +6,8 @@ import com.typesafe.config.{ ConfigFactory, ConfigObject, ConfigValueFactory }
 import org.scalacheck.{ Arbitrary, Gen }
 import pureconfig._
 import pureconfig.error._
-import pureconfig.module.magnolia.reader.auto._
-import pureconfig.module.magnolia.writer.auto._
+import pureconfig.module.magnolia.auto.reader._
+import pureconfig.module.magnolia.auto.writer._
 
 class CoproductConvertersSuite extends BaseSuite {
